backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-socketio==5.12.0
aiohttp==3.11.11
ultralytics==8.3.50
opencv-python==4.10.0.84
numpy==1.26.4
pillow==11.0.0
python-multipart==0.0.20
websocket-client>=1.8.0
networkx==3.2.1

# Test client dependencies
colorama==0.4.6
google-genai>=1.0.0
python-dotenv>=1.0.0
cachetools>=5.0.0
